Move cosmos errors to it's one pkg
