+++ cosmos/errors/errors.go
@@ -1,4 +1,4 @@
-package cosmos
+package errors
 
 import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
